backend/requirements.txt
fastapi
python-multipart
uvicorn
torch
transformers
pillow